修改: 修改拍照模式切换时判断，添加AI状态数据对象定义，添加空中管制事件上报消息实现定义；

+++ cloud-sdk/src/main/java/com/dji/sdk/cloudapi/device/CenterNode.java
@@ -6,7 +6,7 @@ package com.dji.sdk.cloudapi.device;
  */
 public class CenterNode {
 
-    private Integer sdrId;
+    private Long sdrId;
 
     private String sn;
 
@@ -18,11 +18,11 @@ public class CenterNode {
                 '}';
     }
 
-    public Integer getSdrId() {
+    public Long getSdrId() {
         return sdrId;
     }
 
-    public CenterNode setSdrId(Integer sdrId) {
+    public CenterNode setSdrId(Long sdrId) {
         this.sdrId = sdrId;
         return this;
     }
